Move Tile to tile.rs
